Change type of double

+++ evalAssignment1/update_locations.c
@@ -1,20 +1,20 @@
 #include "update_locations.h"
 
-// Create a list of 'size' floating point numbers in the range [-bound, bound]
-float * generate_random_list(int32_t size, int32_t bound) {
-    float* array;
-    array = (float*)malloc(sizeof(float) * size);
+// Create a list of 'size' doubleing point numbers in the range [-bound, bound]
+double * generate_random_list(int32_t size, int32_t bound) {
+    double* array;
+    array = (double*)malloc(sizeof(double) * size);
 
 
     for (int i = 0; i < size; i++) {
-        float num = (float)(rand() % (2*bound+1)) - bound;
+        double num = (double)(rand() % (2*bound+1)) - bound;
         array[i] = num;
     }
     return array;
 }
 
 // Update position by velocity, one time-step
-void update_coords(float * xs, float * ys, float * zs, float * vx, float * vy, float * vz, int32_t size){
+void update_coords(double * xs, double * ys, double * zs, double * vx, double * vy, double * vz, int32_t size){
     for (int i = 0; i < size; i ++) {
         xs[i] = xs[i] + vx[i];
         ys[i] = ys[i] + vy[i];
@@ -22,8 +22,8 @@ void update_coords(float * xs, float * ys, float * zs, float * vx, float * vy, f
     }
 }
 
-float sum(float * array, int32_t size) {
-    float sum = 0;
+double sum(double * array, int32_t size) {
+    double sum = 0;
     for (int i = 0; i < size; i++) { sum += array[i];}
     return sum;
 }
@@ -37,12 +37,12 @@ int main(int argc, char* argv[]) {
     int32_t size = atoi(argv[1]);
     int32_t iters = atoi(argv[2]);
     srand(size);
-    float * xs = generate_random_list(size, 1000);
-    float * ys = generate_random_list(size, 1000);
-    float * zs = generate_random_list(size, 1000);
-    float * vx = generate_random_list(size, 1);
-    float * vy = generate_random_list(size, 1);
-    float * vz = generate_random_list(size, 1);
+    double * xs = generate_random_list(size, 1000);
+    double * ys = generate_random_list(size, 1000);
+    double * zs = generate_random_list(size, 1000);
+    double * vx = generate_random_list(size, 1);
+    double * vy = generate_random_list(size, 1);
+    double * vz = generate_random_list(size, 1);
 
     struct timespec start, stop;
     clock_gettime(CLOCK_MONOTONIC, &start);
@@ -50,7 +50,7 @@ int main(int argc, char* argv[]) {
         update_coords(xs, ys, zs, vx, vy, vz, size);
     }
     clock_gettime(CLOCK_MONOTONIC, &stop);
-    float chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
+    double chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
     double obj_count = (double)size*(double)iters;
     double time_taken;
     time_taken = (stop.tv_sec - start.tv_sec) * 1e9;
